Move table import-source to kart/tabular
