+++ kart/tabular/base_dataset.py
@@ -4,8 +4,8 @@ import logging
 import time
 
 from kart import crs_util
-from kart.import_source import ImportSource
-from kart import meta_items
+from .import_source import ImportSource
+from . import meta_items
 from kart.exceptions import InvalidOperation
 from kart.promisor_utils import LibgitSubcode
 from kart.serialise_util import json_unpack, ensure_text
