+++ kart/tabular/ogr_import_source.py
@@ -10,20 +10,20 @@ import click
 from osgeo import gdal, ogr
 
 
-from . import crs_util
-from .exceptions import (
+from kart import crs_util
+from kart.exceptions import (
     InvalidOperation,
     NotFound,
     NotYetImplemented,
     NO_IMPORT_SOURCE,
     NO_TABLE,
 )
-from .geometry import ogr_to_gpkg_geom
+from kart.geometry import ogr_to_gpkg_geom
 from .import_source import ImportSource
-from .ogr_util import get_type_value_adapter
-from .output_util import dump_json_output
-from .tabular.schema import Schema, ColumnSchema
-from .utils import ungenerator, chunk
+from kart.ogr_util import get_type_value_adapter
+from kart.output_util import dump_json_output
+from .schema import Schema, ColumnSchema
+from kart.utils import ungenerator, chunk
 
 
 # This defines what formats are allowed, as well as mapping
