+++ kart/tabular/sqlalchemy_import_source.py
@@ -6,17 +6,17 @@ import click
 import sqlalchemy
 from sqlalchemy.orm import sessionmaker
 
-from .exceptions import (
+from kart.exceptions import (
     NotFound,
     NotYetImplemented,
     NO_IMPORT_SOURCE,
     NO_TABLE,
 )
 from .import_source import ImportSource
-from .output_util import dump_json_output
-from .tabular.schema import Schema
-from .sqlalchemy import DbType, separate_last_path_part, strip_username_and_password
-from .utils import chunk, ungenerator
+from kart.output_util import dump_json_output
+from .schema import Schema
+from kart.sqlalchemy import DbType, separate_last_path_part, strip_username_and_password
+from kart.utils import chunk, ungenerator
 
 
 class SqlAlchemyImportSource(ImportSource):
